refactor!: do not export internals
