+++ source/collect/CollectService.ts
@@ -2,8 +2,8 @@ import type ts from "typescript";
 import type { ResolvedConfig } from "#config";
 import { Diagnostic, DiagnosticOrigin } from "#diagnostic";
 import { EventEmitter } from "#events";
+import { AbilityLayer } from "#layers";
 import type { ProjectService } from "#project";
-import { AbilityLayer } from "./AbilityLayer.js";
 import { CollectDiagnosticText } from "./CollectDiagnosticText.js";
 import { ExpectNode } from "./ExpectNode.js";
 import { IdentifierLookup, type TestTreeNodeMeta } from "./IdentifierLookup.js";
