+++ source/collect/TestTree.ts
@@ -1,7 +1,7 @@
 import type ts from "typescript";
+import type { SuppressedError } from "#layers";
 import type { ExpectNode } from "./ExpectNode.js";
 import type { TestTreeNode } from "./TestTreeNode.js";
-import type { SuppressedError } from "./types.js";
 import type { WhenNode } from "./WhenNode.js";
 
 export class TestTree {
