+++ source/json/JsonScanner.ts
