+++ source/json/types.ts
@@ -1,0 +1,1 @@
+export type OptionValue = Array<OptionValue> | string | number | boolean | null | undefined;
