+++ source/layers/helpers.ts
@@ -1,0 +1,5 @@
+import type ts from "typescript";
+
+export function nodeIsChildOfExpressionStatement(compiler: typeof ts, node: ts.Node): boolean {
+  return compiler.isExpressionStatement(node.parent);
+}
